This code simulates the soccer seasons
In the league, each team plays each other twice.

So with 20 teams, each team plays 38 matches
A season comprises 380 total matches
The outcome for a season is based upon:

Points scored, where teams get:
3 points for a win
1 point for a draw
0 points for a loss
If two teams have the same number of points, their position is based on the difference between goals scored and goals conceded
If they are equal on points and goal difference, then goals scored are the next tiebreaker.
If they are equal on all of these margins a tie is awarded, but here I will just have you break ties randomly using a uniform draw.
Position Earnings
I lifted the following image from the internet. Whether it's correct or not, we're going to use this as our yardstick for converting league placement into an earnings figure. Don't worry about the categories. However, for every position, you will need to measure a monetary amount (here in GBP).

PremLeagueValue_BarChart.png 

Questions to Answer
Assemble a ranking of the teams from best to worst according to their expected position in the table.
Illustrate this ranking with a clear visualization
Which teams have the most variability in their simulated earnings?
Again, illustrate this variability with a clear visualization.
Which teams stand to benefit the most in monetary terms from a "lucky win" (converting one of their losses to a win)?
Again provide a visualization of this marginal effect
Finally, for each team, calculate the monetary benefit to each team from either (i) an increase of 10% to their expected goals scored (holding everyone else constant), or (ii) a decreasing in their expected goals conceded by 10% (Note that this means 40 separate simulations!)
Use this to provide a clear visualization of the benefits of investing in offense versus defense by team